add image caption tests

+++ test/test_modality.py
@@ -6,41 +6,48 @@ class TestDataModality(unittest.TestCase):
 
     def test_string_pair(self):
         from h5record.dataset import H5Dataset
-        from h5record.attributes import String
+        from h5record.attributes import String, Integer
         schema = {
             'sentence1': String(name='sentence1'),
-            'sentence2': String(name='sentence2')
+            'sentence2': String(name='sentence2'),
+            'label': Integer(name='label')
         }
+
         data = [
-            ['HDF5 supports chunking and compression.', 'You may want to experiment'],
-            ['Starting to load file into an HDF5 file with chunk size','and compression is gzip' ],
-            ['Reading from an HDF5 file which you will probably be','about to overwrite! Override this error only if you know what youre doing '],
+            ['HDF5 supports chunking and compression.', 'You may want to experiment', 0],
+            ['Starting to load file into an HDF5 file with chunk size','and compression is gzip', 1 ],
+            ['Reading from an HDF5 file which you will probably be','about to overwrite! Override this error only if you know what youre doing ', 1],
         ]
 
         def pair_iter():
             for row in data:
                 yield {
                     'sentence1': row[0],
-                    'sentence2': row[1]
+                    'sentence2': row[1],
+                    'label': row[2]
                 }
+        if os.path.exists('question_pair.h5'):
+            os.remove('question_pair.h5')
 
         dataset = H5Dataset(schema, './question_pair.h5', pair_iter())
         for idx in range(len(data)):
             row = dataset[idx]
-            sent1, sent2 = data[idx]
+            sent1, sent2, label = data[idx]
             assert sent1 == row['sentence1']
             assert sent2 == row['sentence2']
-        os.remove('question_pair.h5')
+            assert label == row['label']
 
+        os.remove('question_pair.h5')
 
         dataset = H5Dataset(schema, './question_pair.h5', pair_iter(),
             data_length=len(data), chunk_size=4)
 
         for idx in range(len(data)):
             row = dataset[idx]
-            sent1, sent2 = data[idx]
+            sent1, sent2, label = data[idx]
             assert sent1 == row['sentence1']
             assert sent2 == row['sentence2']
+            assert label == row['label']
 
         os.remove('question_pair.h5')
 
@@ -66,5 +73,20 @@ class TestDataModality(unittest.TestCase):
             'test/images/3.jpeg',
             'test/images/4.jpeg',
         ]
+        def pair_iter():
+            for (caption, img_path) in zip(captions, image_path):
+                yield {
+                    'image': schema['image'].read_image(image_path[0]),
+                    'caption': caption
+                }
+        if os.path.exists('img_caption.h5'):
+            os.remove('img_caption.h5')
 
-        
+        dataset = H5Dataset(schema, './img_caption.h5', pair_iter(),
+            data_length=len(image_path), chunk_size=4)
+
+        for idx in range(len(image_path)):
+            data = dataset[idx]
+            assert data['caption'] == captions[idx]
+
+        os.remove('img_caption.h5')
